pyproject.toml
[project]
name = "aoc"
version = "0.1.0"
authors = [{ name = "Matthew Stibbard" }]
description = "Python solutions to Advent of Code"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "advent-of-code-data>=2.0.4",
    "python-dotenv>=1.0.1",
]
